backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
PyPDF2==3.0.1
python-docx==1.1.0
google-generativeai>=0.7.0
langchain-core==0.2.43
langchain-google-genai==1.0.10
langgraph==0.0.66
websockets==12.0
aiofiles==23.2.1
Pillow==10.1.0
bcrypt==4.1.2